requirements.txt
AWSIoTPythonSDK==1.4.4
tinydb==3.13.0
PyYAML>=6.1
PyJWT==1.7.1
